% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{api_docs}
\alias{api_docs}
\title{Descripción de la api de cammesa
"https://api.cammesa.com/pub-svc/v2/api-docs"}
\usage{
api_docs()
}
\description{
Descripción de la api de cammesa
"https://api.cammesa.com/pub-svc/v2/api-docs"
}
