% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{catalogoPublicaciones}
\alias{catalogoPublicaciones}
\title{El campo NEMO devuelto en la estructura de respuesta de este servicio es
utilizado como identificador de la publicación por el resto de los servicios.}
\usage{
catalogoPublicaciones(vigentes = F)
}
\arguments{
\item{vigentes}{Publicaciones vigentes
"https://api.cammesa.com/pub-svc/public/catalogoPublicaciones"
"https://api.cammesa.com/pub-svc/public/catalogoPublicacionesVigentes"}
}
\value{
data.frame
}
\description{
El campo NEMO devuelto en la estructura de respuesta de este servicio es
utilizado como identificador de la publicación por el resto de los servicios.
}
