% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{findAttachmentByNemoId}
\alias{findAttachmentByNemoId}
\title{Descarga adjuntos por nemo y docId}
\usage{
findAttachmentByNemoId(
  nemo,
  docId,
  attachmentId,
  version = NULL,
  folder = "~\\\\..\\\\OneDrive - Enel Spa\\\\cammesa\\\\downloads",
  file_pattern = function(version, docId, attachmentId) {
     paste0(version, "_",
    docId, "_", attachmentId)
 },
  overwrite = T
)
}
\arguments{
\item{overwrite}{}
}
\value{
list<httr::http_status>
}
\description{
Descarga adjuntos por nemo y docId
}
\examples{
TODO:
nemo puede ser un grupo unificado como DTE_UNIF o PARTE_SEMANAL_UNIF
decidir si va a la carpeta individual o unificada
Prefiero la unificada pero hay que hacer la consulta para identificarlo.
O BIEN : identificar los que son UNIF y probar con todos los nemos
del grupo UNIF (preferible)
usar un WARINING para avisar
}
