% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{findDocumentosByNemo}
\alias{findDocumentosByNemo}
\title{Buscar y opcionalmente descargar documentos de CAMMESA por fecha y NEMO}
\usage{
findDocumentosByNemo(
  nemo = "DTE_EMISION",
  fechadesde = as.POSIXct(trunc(Sys.time() - 30L * 86400L, "month"), tz = tz),
  fechahasta = Sys.time(),
  tz = "America/Buenos_Aires",
  folder = "~\\\\..\\\\OneDrive - Enel Spa\\\\cammesa\\\\downloads",
  download.policy = c("none", "missing.only", "overwrite"),
  queryByMonth = TRUE,
  get.config = list()
)
}
\arguments{
\item{nemo}{nemo de clase de documento de la API de cammesa}

\item{fechadesde}{fecha de consulta desde, fechas desde
"2010-06-01T03:00:00.000Z" fueron probadas}

\item{fechahasta}{fecha de consulta hasta}

\item{folder}{ruta base de descarga (los archivos se descargan en una}

\item{download.policy}{no bajar, sobreescribir o bajar sólo faltantes}

\item{queryByMonth}{divide la consulta en meses}

\item{get.config}{pasado a httr::GET para la descarga de arhcivos}
}
\value{
data.frame de documentos encontrados para las fechas y nemo
especificados (devuelto por la llamada REST a
\code{findDocumentosByNemoRango})
}
\description{
Buscar y opcionalmente descargar documentos de CAMMESA por fecha y NEMO
}
\details{
El período de consultas es dividido en tramos mensuales para la
consulta y luego se concatenan. Por experiencia eso asegura que la consulta
devuelva todos los archivos
}
