% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{obtieneFechaUltimoDocumento}
\alias{obtieneFechaUltimoDocumento}
\title{Fecha último documento
Api REST: "https://api.cammesa.com/pub-svc/public/obtieneFechaUltimoDocumento"}
\usage{
obtieneFechaUltimoDocumento(nemo = "DTE_UNIF")
}
\arguments{
\item{nemo}{}
}
\description{
Fecha último documento
Api REST: "https://api.cammesa.com/pub-svc/public/obtieneFechaUltimoDocumento"
}
