% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{querydt}
\alias{querydt}
\alias{getQueryDT}
\alias{updateQueryDT}
\title{Mantiene una tabla histórica de documentos consultados en un directorio local}
\usage{
getQueryDT(file = "~/../OneDrive - Enel Spa/cammesa/Downloads/query.csv")

updateQueryDT(
  x,
  file = "~/../OneDrive - Enel Spa/cammesa/Downloads/query.csv",
  force = FALSE
)
}
\arguments{
\item{file}{}

\item{x}{datos para actualizar la tabla, normalmente de
findDocumentosByNemo}

\item{force}{}
}
\description{
Mantiene una tabla histórica de documentos consultados en un directorio local
}
\details{
Esta función es provista para ayudar a mantener el histórico de consultas
a la API. Se ha comprobado que en ocasiones los archivos consultados
más viejos han dejado de estar públicos.
}
